requirements.txt
fastapi
uvicorn
python-docx
reportlab
openai
